+++ wom/models/__init__.py
@@ -102,6 +102,7 @@ __all__ = (
     "SnapshotData",
     "Snapshot",
     "SnapshotTimelineEntry",
+    "SocialLinks",
     "Team",
     "Top5ProgressResult",
 )