+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add `at_random` method to `BaseEnum` for generating an enum variant at random.
 - Add `levels` property to `CompetitionParticipationDetail`.
+- Add `patron`, `banner_image`, and `profile_image` properties to `Group`.
+- Add `SocialLinks` model and `social_links` property to `GroupDetail`.
 
 ## Changes